Stored num_sites/num_split [demos] [skip hazardlib]
Former-commit-id: a2f79f5c78b6f24dc54d83b23ee8779ffdb8d7fe [formerly 5f878d58d66f3a10f2bc112cf55648aa096ed737]
Former-commit-id: b87524a750ed813dbc1e98e31118d0587072ad22

+++ openquake/commonlib/source.py
@@ -959,7 +959,7 @@ class SourceInfo(object):
         ('num_ruptures', numpy.uint32),    # 2
         ('calc_time', numpy.float32),      # 3
         ('split_time', numpy.float32),     # 4
-        ('num_sites', numpy.uint32),       # 5
+        ('num_sites', numpy.float32),      # 5
         ('num_split',  numpy.uint32),      # 6
         ('events', numpy.uint32),          # 7
     ])
